Made primitive columns fully polymorphic with the new TypeMapper trait.
You can declare a table column of any type T and lift any value of type
T to a ConstColumn[T] as long as an implicit TypeMapper[T] is in scope.

Operators on specific column types (currently only used for Boolean) are
implemented in a sub-trait of ColumnOps which can either be instantiated
with an implicit conversion or mixed into a column class (as done for the
boolean operators, to prevent some unnecessary conversions).