Eliminate Pure(TableRef(_)) in FROM clauses by inlining.
This is necessary because SQL comprehensions are not true monad
comprehensions: They do not allow sub-queries in FROM clauses to
reference tables previously introduced in other FROM clauses of the
same query.

Not all TableRefs can be eliminated that way so we check for this
situation and throw an SQueryException if it occurs. A more elaborate
transformation could handle some more cases.