Make Query invariant in U.
Previously, we got away with Query being covariant because you always
needed a Query[ColumnBase[T]] to do something sensible and ColumnBase
is invariant. With the new Packing-based approach, Query itself should
(and can) also be invariant. Making it invariant in E would be much harder
and it seems unnecessary.