Recursively push refs through ExpandedTableRefs.
Expanding Refs to TableNodes into Products was not enough. There can be
references by FieldSymbol into such a Ref, so we now expand them to
ExpandedTableRefs, similar to the expansion of TableNodes to
ExpandedTables. When a Select(_, FieldSymbol) targets a reference to
such an ExpandedTableRef, we can follow it recursively to the original
ExpandedTable and then rewrite the expansions all the way up.