Introduce standard library of function and operator Symbols.
Specialized nodes for these operations are replaced by a new generic
Apply node that takes a Symbol for the operation and a list of child
Nodes as its operands.

We also need to introduce the notion of TypedNodes (and more generally,
Typed entities other than Columns) to preserve the necessary type
information for the drivers.