Don't expand refs if the expansion would be dropped later.
We used to expand all refs in Columnizer and then decide in PathRewriter
whether to keep the ref or the expansion (depending on the position of
the expansion) but this only works when the expansion is actually a
TableRefExpansion node. When a path points to a ProductNode of two
TableRefs, for example, the expansion is another ProductNode of two
TableRefExpansions, and the information that it used to be a single Ref
is lost. PathRewriter cannot deal with such situations in a sensible way,
so we prevent creating the expansion in the first place. The downside is
more complexity in Columnizer because it has to be able to deal with
unexpanded refs in later stages.