Remove mutable names from AnonSymbols.
They can interfere with reuse of symbols in a different thread or even
in the same thread (because it is easy to create clashing names). Instead,
we keep name assignments in a SymbolNamer which is part of the compiler
state and also gets passed to the code generator in the end.