Remove TypeMapper abstraction.
TypeMapper becomes ast.TypedType (plus BaseTypedType, OptionTypedType and
NumericType) but it does not support the TypeMapperDelegate lookup any
more. Instead, both ast.StaticType and JdbcType (the former
TypeMapperDelegate) implement TypedType. Implicits for StaticType can be
used in the lifted embedding, AST and compiler. Implicits for the
JdbcTypes are provided by the driver for use in user code. The driver also
knows how to map the few remaining StaticTypes to the right JdbcType.

This removes the dependency on JDBC-specific types from the lifted
embedding. Profiles other then JdbcProfile can provide their own set of
types along with their own type implementation.

The downside of this change is that there is no mechanism to prevent users
from using types from a different driver (even if it implements an
unrelated profile). This will lead to subtle runtime errors if the
types have compatible implementations (e.g. they are all JdbcTypes) and
to runtime exceptions during type lookup when the implementations don't
match. But the SimpleQL imports make it very easy to use the right types
and it requires some work to use the wrong ones.