Add codeGen phase to generate SQL code within the AST.
This allows us to build more complex trees that contain client-side
mappings on top of SQL. Later on we should be able to use this as a basis
for nested collections and distributed queries, both of which require
similar mechanisms.

Concrete instances of codeGen are created by the drivers and appended
to the driver's common QueryCompiler configuration to obtain different
compilers for the different kinds of statements (select, update, delete)
that can be built this way.

This initial commit does not generate client-side mappings yet because
not all uses of the query AST and the old code generator model have been
adapted to use the codeGen phase.