Hoist client ops up into ResultSetMapping and make that fully typed.
Client ops were not hoisted out of the outer Comprehension before because
that only works when returning a StructNode (so we can just work with
unique symbols instead of having to track paths and scopes). Now we
temporarily change the outer Comprehension to return a StructNode,
wrap it in a ResultSetMapping, perform the hoisting, then rewrite back
to ProductNode form.

Node.nodeWithComputedType gets a new flag for incremental instead of
full retyping (i.e. only nodes with UnassignedType are changed). This
makes it very easy to compute the types at the ResultSetMapping level
(where everything except the ResultSetMapping itself and its client-side
projection is already typed) in an efficient way.