Force comprehension inlining to avoid illegal SQL comprehensions.
When a sub-comprehension refers to a previous generator from an outer
comprehension, force inlining even in the presence of non-trivial
projections. Usually we don't want to inline projections that do more
than just aliasing to avoid code duplication but it's necessary to avoid
illegal SQL comprehension shapes.

Fixes issue #83. Test in NewQuerySemanticsTest.testAdvancedFusion.