Remove dependency on linearizers from constraint handling and templates.
Index/constraint handling used to rely on linearizers for getting a list
of referenced columns. Instead we now use the Shape to pack() the original
value, grab the packed Node and extract the linear sequence from that
(which can be done independently of the Shape -- see
ExtraUtil.linearizeFieldRefs).

QueryTemplates not only do not need the linearizers any more, they can
even be unified with QueryInvokers (so that a QueryInvoker[P] is simply a
QueryTemplate[Unit, P]) with the additional unit invoker mix-ins.