fixes sessions leaking withSession in some cases
BaseSession.conn was a lazy val. If conn was not materialized within the
withSession scope the "finally s.close()" call did not do anything. If the
Session object was leaked as a return value, state of a closure, a future
or else, which materialized conn, a connection was opened but never
closed.