Move QueryTemplate functionality up into BasicProfile.
BasicProfile now defines abstract types for parameterized and applied
queries that are pre-compiled by the driver, along with a standard
implementation.

Actually executing these queries is done through new implicit conversions
to QueryInvoker and QueryExecutor. This change makes the execution of
plain queries and parameterized queries symmetrical and allows the latter
to be defined in a profile- and driver-independent way.