Move Table and Sequence into RelationalProfile, DDL into BasicProfile.
DDL is abstracted into BasicProfile.SchemaDescription which provides
concatenation but no internal representation of schemas. The string
representation is added by the old DDL type which has been moved into
SqlProfile.

With DDL being available at that level, overriding Table in JdbcProfile
is no longer necessary, so the complete definition is now at the level of
RelationalProfile. Sequence is used in a similar way to Table, so we also
make it part of the cake at the RelationalProfile level.