Move more functionality up to Basic/Relational/SqlProfile.
An abstract SchemaDefinition type (for DDLs) along with DDLInvoker goes
into BasicProfile (which is debatable - we may want to split that up later
into a really basic profile which is only for querying and one that
provides write access to the data)..

ColumnType and abstract implicits for the standard Scala types go into
RelationalProfile. We can move it further up to BasicProfile later when
we have a clearer picture how non-tabular data structures should be
represented.

All QueryExecutor functionality goes into BasicProfile. I'm also removing
the special ToQueryExecutor workaround for SI-3346. Our standard .shaped
workaround is enough to cover this case.

The QueryCompilers for querying and inserting are also moved up to
BasicProfile.