Always add parentheses around subqueries in a SELECT clause.
Note that this doesn't catch the cases where the expr() call at the end
falls through to a subquery which is introduced on the fly by
convertToComprehension(). We cannot handle these cases in a compatible
way in 1.0. With the typed ASTs in 2.0, it should be possible to perform
all these conversions before the code generator in an extra compiler
phase so that this fix will work for them as well.

Fixes issue #101. Test case in NewQuerySemanticsTest.testSubquery.