Throw an exception when trying to lift a bound table row to a Query.
This is almost never what you want but it can happen accidentally when
you try to think in SQL semantics but have to build a query with
collection semantics. See issue #101 for a motivating example. You can
still perform this conversion explicitly if needed. Unfortunately, there
is no way to detect this situation at compile-time (barring a redesign
of the table objects).