Support schema names for tables in all statements except INSERTs.
The InsertBuilder API needs to change in a binary-incompatible way so we
cannot do this last bit in 1.0.x. The new quoteTableName method belongs
into BasicSQLUtilsComponent next to quoteIdentifier. It is placed as a
private method into BasicStatementBuilderComponent to ensure binary
compatibility with Slick 1.0.0. There should be a TableNode.schemaName
method but we cannot add that, either, without breaking binary
compatibility, so we pattern-match in quoteTableName instead. This needs
to be cleaned up after merging 1.0 back into master.

Test case in SchemaSupportTest. Ideally, all TestKit tests would use
schema names, but that will require some major work on TestKit.

Fixes issue #19.