Inline lifted count(*) aggregates over scalar Comprehensions.
count(*) in a sub-Comprehension is usually turned into a count(1) but if
the sub-Comprehension is known to be scalar, it can be lifted out and
inlined as a constant 1.

We do this mainly for the benefit of MS Access (which does not support
the subquery in this case) but we can safely run it as part of the
standard query compiler.

Fixes ticket 230. No extra test (NewQuerySemanticsTest.testSubQuery was
broken on Access).