Expand symbols in the "by" clause of a GroupBy node.
A proper fix would entail removing the "by" Symbol from GroupBy (but we
cannot do that in 1.0.x due to binary compatibility constraints). It
should not be needed anymore in the modern Slick AST. ExpandRefs
incorrectly assumed that only "from" clauses would have symbols so we add
a special case for GroupBy for now.

Fixes issue #141. Test in AggregateTest.groupBy (q5).