Improve threadLocalSession API
threadLocalSession is renamed to dynamicSession to be consistent with
Scala terminology (Java calls it thread-local).

The withSession/withTransaction methods that make use of dynamicSession
are renamed to withDynSession/withDynTransaction to avoid overloading
the standard withSession/withTransaction methods. This makes it possible
to use the standard methods without explicit type annotations, e.g.

  db withSession { implicit session => ... }

instead of

  db withSession { implicit session: Session => ... }

Fixes issue #117.