Use the ++ operator for unionAll.
Since the default for SQL queries is to allow multiple occurences of
rows and this is well represented in Slick by using Seq as the default
type constructor in lifted.Query, the standard ++ operator for Scala
collections should be mapped to unionAll. Later, when we add support
for other collection types, the semantics will depend on the collection
type and we can remove the separate union and unionAll methods.