Move the conversion of values to Nodes into Shape.
Node.apply had a hard-coded conversion for Product. This needs to be
handled by the Shape if we want to support custom Shapes for other
container types.

We can also disentangle Node from NodeGenerator. Multi-level indirection
is no longer necessary with the clear separation between the AST and the
lifted embedding that we have now.