Use fixed version numbers for fmpp's dependencies.
This requires disabling transitive dependencies and specifying all of
them manually with the version that would currently be resolved from
Maven Central. All dependencies are specified as version ranges in fmpp
which means that we don't get a reproducible build or the ability to
start the sbt project without an Internet connection unless we exclude
them and do it correctly.