Fix fusion of Take and Drop operators.
This got broken somewhere along the way during the rewrite of the
central phases of the query compiler and the ensuing changes to
convertToComprehensions. Conversion is inside-out now instead of
outside-in, so the TakeDrop extractor needs to take already converted
operators into account.

This fixes PagingTest on SQL Server which was broken since the original
changes. (The fact that this had any influence on that test shows just
how brittle the code generator is when it comes to correctly propagating
orderings. It's basically still a kludge that relies on fusion doing a
good enough job.)