Fix links in documentation.
Some method links in connection.rst still don't light up in scaladoc
(but at least the pages are now correct). They are probably wrong but
guessing the right ones is a nightmare. Scaladoc needs to add a way to
get proper links out of the generated pages to make this usable.