Infer correct NominalType views even without the NominalType.
In inferTypes, we used to replace structural views of NominalTypes after
constructing all table types. At that point, however, type inference
(which runs before table types can be determined) may have copied a
structural view on its own (without the enclosing NominalType) into
another type. For instance, this happens in the typing of GroupBy, where
the "by" type needs to be copied purely as a structural view. In these
cases the resulting NoType could not be patched to the correct type. To
fix this problem we use the new type UnassignedStructuralType(ts) with
the same TypeSymbol as the surrounding NominalType in table types, which
can be patched even after the NominalType has been lost.

The only remaining use of NoType in StarAnd can be replaced by
UnassignedType, thus eliminating NoType entirely.

Tests in AggregateTest.testGroupBy. Fixes issue #252.