The experimental patch to leave out parenthesis around the Union.
Let's say a example query.

  tabA.union(tabB).union(tabC)

this patch creates sql like below

  select ... from tA union select ... from tB union select ... from tC

original implementation creates below sql.

  select ... from (select ... from (select ... from tA union select ... from tB) union select ... from tC)

this patch is intended to solve the With Clause's parenthesis probelm
because there are some syntax restrictions like ORA-32042.

This patch doesn't use the Union in node permutation(fuse).
the Union is only used with Query.union,
and all union-ed nodes are collected to Comprehension's unionNodes as wrapper node InternalUnionNode.
so, there is no the Union in JdbcStatementBuilderComponent phase.

Problem
1:
With this patch, in the Derby DBMS, NewQuerySemanticsTest.scala:350 is not executed.
Otherwise, there is no problem in another DBMSs.
so, I think that maybe this is the Derby's bug.
As workaround, I skipped the line of the test only when the Derby DBMS is used.
Please check this.

Thanks.