Generate table of driver capabilities for the user manual.
A simple command line app `BuildCapabilitiesTable` in TestKit builds the
actual table, which is then included with the `csv-table` directive in
the manual. We can easily link to the scaladocs from the table because
the CSV file has access to the Sphinx extensions as the regular RST
sources.

Fixes issue #435.