Try for the Select Distinct.
Hi.

This work is for the Select Distinct.
In the Slick, in order to use the select distinct, below syntax is used.
  qA.groupBy(t=>t).map(_._1)
And, result sql is like below.
  select x1.colA, x1.colB from (...) x1 group by x1.colA, x1.colB

In this work, below syntax is possible, also.
  qA.distinct
And, result sql is like below.
  select distinct x1.colA, x1.colB from (...) x1

This doesn't introduce any Node classes. Instead, add a one flag into Comprehension.
And, the .distinct method is just a macro for groupBy(t=>t).map(_._1).

Thanks.