Remove BoneCP support and drop configuration compatibility with Play.
With the modularization in Play and the needs of Reactive Slick, the
goal is to use Slick’s own pool configuration in Play, thus trading
configuration compatibility in Play across different database frameworks
for consistency of database configurations within the same framework
across different kinds of applications (Play and non-Play).

We keep some old config key names where the corresponding HikariCP
option names are unnecessarily verbose. Both naming styles are supported
for both, HikariCP and the non-pooled DriverJdbcDataSource.

Old (already available in Slick 2.0) keys which are no longer officially
supported, and their replacements:
  autocommit     -> autoCommit
  isolation      -> transactionIsolation
  defaultCatalog -> catalog
For compatibility with 2.0, these keys are still accepted for non-pooled
configurations but they trigger a log warning stating that the key is
deprecated.