Remove the option to disable auto-commit for new connections.
Slick’s internal transaction handling requires all new connections to be
in auto-commit mode, so there is no reason to disable auto-commit. This
is also the default according to the JDBC specification. (Connection
pools usually have an option to disable auto-commit but the config keys
do not apply to connections created on top of an externally managed
DataSource anyway).

We also remove HikariCP’s `isolateInternalQueries` options which is
meaningless in auto-commit mode.

Using the old `autocommit` config key (which was already supported in
Slick 2.0) now triggers a log error stating that it is no longer
supported.