Add concurrency stress test for streaming API:
https://travis-ci.org/slick/slick/builds/44488287 shows a test failure
of createPublisher1MustProduceAStreamOfExactly1Element for JdbcBackend:

Publisher (...) produced no elements expected [true] but found [false]

In response to this failure I added a concurrency stress test in the
hope of being able to reproduce it. `StreamsStressTest` runs the same
scenario as the failed test, except many times over and in parallel. On
my MBP I ran into issues with H2 on JDK 8 with as little as 2 threads
and a few hundred iterations on each, which looks like a bug in either
H2 or the version of JDK 8 I have installed. Local tests with up to 16
threads and 10000 iterations on each thread ran successfully against
H2 on JDK 6 and 7, and against Derby on JDK 6, 7 and 8.