Use force-insert semantics for inserting from queries and expressions.
It always worked this way up to Slick 2.0. In 2.1 we made soft inserts
the default but they are not used for inserting from queries and
expressions. The implementation in 2.1 had two problems: The generated
statements were wrong (using the force-insert statement instead of the
soft-insert statement) and the method names didn’t reflect the semantics
properly.

Fixes #983. Test in InsertTest.testForced.