Package structure and project layout refactoring:
The main goal is to move everything from `scala.slick` to `slick`.
When the project started in 2011, the goal was to eventually make it a
part of the Scala standard library. This was abandoned long ago with
the modularization of the standard library, so we’re bringing Slick in
line with Play and Akka, both of which also use a top-level package.

Fixing problems with existing package names:
- scala.slick.collection.heterogenous -> slick.collection.heterogeneous
- All doc snippets in com.typesafe.slick.docs
- All doc snippet tests in com.typesafe.slick.docs.test

Except for the doc snippets, all source files are in the proper
subdirectory for their package.

The package object for scala.slick provides deprecated aliases for many
important types and values.

Since most files get moved around anyway, this is a good opportunity to
clean up the project layout: The main project `slick` finally gets its
own subdirectory next to the other projects.