Lazily instantiate the ExecutionContext in AsyncExecutor:
This is required in 3.0 because an AsyncExecutor is created for every
Database and there was no need (and no way) to shut down Databases in
earlier versions of Slick. The migration guide advises that close() is
not strictly necessary when only using the old API in 3.0, so we may
not create a thread pool until it is needed (by the new API).