Distinguish between high- and low-priority jobs in AsyncExecutor:
When using a limited queue size, jobs can get rejected if the queue is
full. You want this to happen for new DBIO executions but not for
continuations of existing ones (e.g. finishing a transaction that was
started earlier and then interrupted for an asynchronous computation).

We now create twice the specified queue size and allow the extra size to
be used for high-priority jobs (which have to implement
AsyncExecutor.PrioritizedRunnable). The first SynchronousDatabaseAction
being executed as part of a DBIOAction is submitted with low priority,
subsequent actions and stream continuations with high priority. Note
that the priority does not change the position in the queue. It only
determines whether the job may be enqueued at this time.

ManagedArrayBlockingQueue, a copy of j.u.c.ArrayBlockingQueue,
implements this behavior. Unfortunately it is not possible to extend
ArrayBlockingQueue in order to do this because some essential internals
(like the lock) are not visible to subclasses.

Fixes #1097.